+++ package.json
@@ -1,10 +1,10 @@
 {
-  "name": "basicnode",
+  "name": "air_liquide_test",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "server.js",
   "scripts": {
-    "start": "node airLiquide.js"
+    "start": "node server.js"
   },
   "author": "",
   "license": "ISC",
