--- views/index.html
@@ -1,14 +1,0 @@
-<html>
-
-<head>
-    <title>Test</title>
-    <link rel="stylesheet" type="text/css" href="/css/style.css">
-</head>
-
-<p id="container">Hello Michael!</p>
-
-<body>
-    <script src="app.js"></script>
-</body>
-
-</html>
